Move events to individual cogs files
